+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0-SNAPSHOT - 03/04/17
+- Fix missing ParseObject class name
 - Added GWT Marshaller and Unmarshaller for ParseObject
 
 ## 0-SNAPSHOT - 02/24/17